+++ Demos/AllOverIt/Caching/GenericCacheDemo/Keys/PropKeyByBinding.cs
@@ -3,7 +3,7 @@ using AllOverIt.Reflection;
 
 namespace GenericCacheDemo.Keys
 {
-    internal sealed class PropKeyByBinding : GenericCacheKey<BindingOptions>
+    internal sealed record PropKeyByBinding : GenericCacheKey<BindingOptions>
     {
         public PropKeyByBinding(BindingOptions bindingOptions)
             : base(bindingOptions)