+++ Change Log.md
@@ -45,6 +45,7 @@
 * Added ToMemoryStream() extension method for string.
 * Renamed IDictionary extension method from Concat() to Combine() to avoid namespace clashes
 * Renamed IList extension method from AddRange() to AddMany() to avoid namespace clashes
+* Changed generic cache keys to be records
 
 
 ## AllOverIt.AspNetCore